+++ projects/plugins/jetpack/extensions/blocks/blogroll/index.js
@@ -1,3 +1,4 @@
+import { InnerBlocks } from '@wordpress/block-editor';
 import { __ } from '@wordpress/i18n';
 import { getIconColor } from '../../shared/block-icons';
 import attributes from './attributes';
@@ -60,7 +61,7 @@ export const settings = {
 		},
 	},
 	edit,
-	save: () => null,
+	save: () => <InnerBlocks.Content />,
 	attributes,
 	providesContext: {
 		showAvatar: 'show_avatar',